Added global exceptions with custom exception classes ResourceNotFoundException and BadRequestException. Changed controllers and services accordingly. Added title checking(if title is not null or empty). Added ExceptionResponseDTO

+++ src/main/java/com/example/wishes/service/WishesService.java
@@ -1,5 +1,7 @@
 package com.example.wishes.service;
 
+import com.example.wishes.exception.BadRequestException;
+import com.example.wishes.exception.ResourceNotFoundException;
 import com.example.wishes.model.Wish;
 import com.example.wishes.model.WishList;
 import com.example.wishes.repository.WishListRepository;
@@ -18,51 +20,71 @@ public class WishesService {
     @Autowired
     private WishRepository wishRepository;
 
+    //helper methods : Exception handling
+    private void validateTitle(String title, String entityName) {
+        if (title == null || title.trim().isEmpty()) {
+            throw new BadRequestException(entityName + " title cannot be empty");
+        }
+    }
+
+    private Wish getWishAndValidateWishList(int wishId, WishList wishList){
+        Wish wish = wishRepository.findById(wishId)
+                .orElseThrow(()-> new ResourceNotFoundException("Wish not found on id: " + wishId));
+        if(wish.getWishList().getId() != wishList.getId()){
+            throw new BadRequestException("Wish id: " + wishId + " is not from id: "+ wishList.getId() + " wish list");
+        }
+        return wish;
+    }
+
+
 
     //get all wish lists
     public List<WishList> getAllWishLists() {
+
         return wishListRepository.findAll();
     }
 
     //get wish list by id
     public WishList getWishListByID(int id){
-        return wishListRepository.findById(id).orElse(null);
+
+        return wishListRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Wish List not found with id : " + id));
     }
 
     //create a wish list
     public WishList createWishList(String title){
+        validateTitle(title, "Wish List");
         return wishListRepository.save(new WishList(title));
     }
 
     //update wish list
     public WishList updateWishList(int idToUpdate, String newTitle){
-        WishList wishListToUpdate = wishListRepository.findById(idToUpdate).orElse(null);
-        if (wishListToUpdate != null) {
-            wishListToUpdate.setTitle(newTitle);
-            return wishListRepository.save(wishListToUpdate);
-        }
-        return null;
+        validateTitle(newTitle, "Wish List");
+        WishList wishListToUpdate = getWishListByID(idToUpdate);
+        wishListToUpdate.setTitle(newTitle);
+        return wishListRepository.save(wishListToUpdate);
     }
 
     //delete wish list
     public void deleteWishList(int id){
+        if(!wishListRepository.existsById(id)){
+            throw new ResourceNotFoundException("Wish List not found on id : " + id);
+        }
         wishListRepository.deleteById(id);
     }
 
 
     //get all wishes in a specific wish list
     public List<Wish> getAllWishes(int wishListId){
-        WishList wishList = wishListRepository.findById(wishListId)
-                .orElseThrow(() -> new RuntimeException("Wish List not found"));;
+        WishList wishList = getWishListByID(wishListId);
         return wishList.getWishes();
     }
 
     //create a wish inside a specific wish list
     public Wish createWish(int wishListId, String title, Boolean isComplete){
-        WishList wishList = wishListRepository.findById(wishListId).orElse(null);
-        if(wishList == null){
-            throw new RuntimeException("Wish List not found");
-        }
+        validateTitle(title, "Wish");
+
+        WishList wishList = getWishListByID(wishListId);
 
         Wish wish = new Wish(title, isComplete);
         wish.setWishList(wishList);
@@ -71,14 +93,9 @@ public class WishesService {
 
     //update a wish inside a specific wish list
     public Wish updateWish(int wishListId, int wishId, String newTitle, boolean isCompleted) {
-        WishList wishList = wishListRepository.findById(wishListId).orElse(null);
-        if (wishList == null) {
-            throw new RuntimeException("Wish List not found");
-        }
-        Wish wishToUpdate = wishRepository.findById(wishId).orElse(null);
-        if (wishToUpdate == null) {
-            throw new RuntimeException("Wish not found");
-        }
+        validateTitle(newTitle, "Wish");
+        WishList wishList = getWishListByID(wishListId);
+        Wish wishToUpdate = getWishAndValidateWishList(wishId, wishList);
 
         wishToUpdate.setTitle(newTitle);
         wishToUpdate.setCompleted(isCompleted);
@@ -88,14 +105,8 @@ public class WishesService {
 
     //mark a wish completed
     public Wish markWishCompleted(int wishListId, int wishId, boolean isCompleted) {
-        WishList wishList = wishListRepository.findById(wishListId).orElse(null);
-        if (wishList == null) {
-            throw new RuntimeException("Wish List not found");
-        }
-        Wish wishToUpdate = wishRepository.findById(wishId).orElse(null);
-        if (wishToUpdate == null) {
-            throw new RuntimeException("Wish not found");
-        }
+        WishList wishList = getWishListByID(wishListId);
+        Wish wishToUpdate = getWishAndValidateWishList(wishId, wishList);
 
         wishToUpdate.setCompleted(isCompleted);
 
@@ -104,11 +115,9 @@ public class WishesService {
 
     // Delete a specific To-Do
     public void deleteWish(int wishListId, int wishId) {
-        WishList wishList = wishListRepository.findById(wishListId).orElse(null);
-        if (wishList == null) {
-            throw new RuntimeException("Wish List not found");
-        }
-        wishRepository.deleteById(wishId);
+        WishList wishList = getWishListByID(wishListId);
+        Wish wish = getWishAndValidateWishList(wishId, wishList);
+        wishRepository.delete(wish);
     }
 
 }
